Games are assigned in the client handler.
More than two clients can connect with the server, they are assigned to
correct new games. Games are started from the handler, so the handler
can keep a reference to the game and pass that to the server.
Little bug fixes: if a client types a wrong IP address he/she is given a
new chance to give a correct IP address. (Indefinitely.) Server shows
when waiting for a client.